+++ packages/terra-framework-docs/src/terra-dev-site/doc/compact-interactive-list/Examples.2/CellContent.jsx
@@ -1,14 +1,16 @@
-import React, { useState, useCallback } from 'react';
+import React, { useState, useCallback, useRef } from 'react';
 import CompactInteractiveList, { alignTypes } from 'terra-compact-interactive-list';
 import {
   IconFeaturedOff, IconFeatured, IconMultipleResultsNormal, IconMultipleResultsNotNormal, IconMultipleResultsCritical,
 } from 'terra-icon';
 import Button from 'terra-button';
 import Hyperlink from 'terra-hyperlink';
+import Checkbox, { CheckboxField } from 'terra-form-checkbox';
+import Popup from 'terra-popup';
 import BasicMenu from '../../menu/example/BasicMenu';
 
 // eslint-disable-next-line no-alert
-const buttonCell = <Button text="Learn more" />;
+const buttonCell = <Button text="Learn more" onClick={() => alert('Learn more button was clicked')} />;
 const anchorCell = <Hyperlink href="https://www.cerner.com" text="Documents" />;
 const iconResultsNormal = <IconMultipleResultsNormal a11yLabel="Results normal" height="1.5em" width="1.5em" />;
 const iconResultsNotNormal = <IconMultipleResultsNotNormal a11yLabel="Results not normal" height="1.5em" width="1.5em" />;
@@ -34,7 +36,56 @@ const FeaturedIcon = () => {
   const [isFeatured, setIsFeatured] = useState(false);
   const onButtonClick = () => setIsFeatured(!isFeatured);
   return (
-    isFeatured ? <Button variant="utility" text="Unfavorite item" icon={<IconFeatured />} onClick={onButtonClick} /> : <Button variant="utility" text="Favorite item" icon={<IconFeaturedOff />} onClick={onButtonClick} />
+    isFeatured
+      ? <Button variant="utility" text="Unfavorite item" icon={<IconFeatured />} onClick={onButtonClick} />
+      : <Button variant="utility" text="Favorite item" icon={<IconFeaturedOff />} onClick={onButtonClick} />
+  );
+};
+
+const PopupWithInteractiveContent = () => {
+  const [isOpen, setIsOpen] = useState(false);
+  const [selected, setSelected] = useState(null);
+  const buttonRef = useRef(null);
+
+  const setButtonNode = useCallback((node) => { buttonRef.current = node; }, []);
+  const getButtonNode = useCallback(() => buttonRef.current, []);
+  const handleButtonClick = useCallback(() => { setIsOpen(true); }, []);
+  const handleRequestClose = useCallback(() => { setIsOpen(false); }, []);
+  const closeAndSetFocus = useCallback(() => {
+    handleRequestClose();
+    buttonRef.current.focus();
+  }, [handleRequestClose]);
+
+  const handleOnChange = useCallback((e) => {
+    const selectedAnswers = [...selected];
+    if (e.currentTarget.checked) {
+      selectedAnswers.push(e.currentTarget.value);
+    } else {
+      selectedAnswers.splice(selectedAnswers.indexOf(e.currentTarget.value), 1);
+    }
+    setSelected(selectedAnswers);
+  }, [selected]);
+
+  return (
+    <>
+      <Button text="Clinical Suite" onClick={handleButtonClick} refCallback={setButtonNode} />
+      <Popup
+        isOpen={isOpen}
+        targetRef={getButtonNode}
+        onRequestClose={handleRequestClose}
+        contentHeight="auto"
+      >
+        {/* eslint-disable-next-line react/forbid-dom-props */}
+        <div style={{ padding: '1em' }}>
+          <CheckboxField legend="Do you want to use any of our clinical applications?">
+            <Checkbox id="drug-database" name="applications[]" labelText="Drug Database" onChange={handleOnChange} value="drug-database" />
+            <Checkbox id="vitals-collection" name="applications[]" labelText="Vitals Collection" onChange={handleOnChange} value="vitals-collection" />
+            <Checkbox id="immunization-manager" name="applications[]" labelText="Immunization Manager" onChange={handleOnChange} value="immunization-manager" />
+          </CheckboxField>
+          <Button text="Submit Request" onClick={closeAndSetFocus} />
+        </div>
+      </Popup>
+    </>
   );
 };
 
@@ -52,7 +103,7 @@ const rows = [
     id: 'row_2',
     cells: [
       { content: iconResultsNormal },
-      { content: 'Initial observation Care/Day High Severity 99220 (2)' },
+      { content: 'Initial Observation Care/Day High Severity 99220 (2)' },
       { content: buttonCell },
       { content: <FeaturedIcon /> },
     ],
@@ -71,7 +122,7 @@ const rows = [
     cells: [
       { content: ' ' },
       { content: 'Sbsq Observation Care/Day High Severity 99226 (4)' },
-      { content: buttonCell },
+      { content: <PopupWithInteractiveContent /> },
       { content: <FeaturedIcon /> },
     ],
   },
@@ -106,7 +157,7 @@ const cols = [
   {
     id: 'Column-2',
     displayName: 'Details',
-    width: '210px',
+    width: '225px',
     align: alignTypes.RIGHT,
     isSelectable: true,
   },