+++ packages/terra-popup/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Changed
+  * Added event handlers to call `stopPropagation()` on key and focus events on PopupContent in other components.
+
 ## 6.80.0 - (April 4, 2024)
 
 * Changed